backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
motor==3.6.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
openai==1.58.1